@@ -1,19 +1,21 @@
 module moqueries.org/cli
 
-go 1.21
+go 1.24.0
+
+toolchain go1.24.3
 
 require (
 	github.com/dave/dst v0.27.3
-	github.com/spf13/cobra v1.8.1
-	golang.org/x/mod v0.20.0
-	golang.org/x/text v0.17.0
-	golang.org/x/tools v0.24.0
-	moqueries.org/runtime v0.4.0
+	github.com/spf13/cobra v1.9.1
+	golang.org/x/mod v0.24.0
+	golang.org/x/text v0.23.0
+	golang.org/x/tools v0.31.0
+	moqueries.org/runtime v0.5.0
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sync v0.8.0 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
+	golang.org/x/sync v0.12.0 // indirect
 	moqueries.org/deephash v0.26.0 // indirect
 )
